+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- [Dropbox] Batch correctly overwrite existing files
+- [Dropbox] Batch now rejects the promise if one writing action failed
+
 ## [2.0.0] - 2017-11-25
 ### Changed
 - GitHub batch fixes and optimization
@@ -51,6 +56,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
   - Total rework of the philosphy
 
-[2.0.0]: https://github.com/silexlabs/unifile/compare/v1.2.0...v
+[Unreleased]: https://github.com/silexlabs/unifile/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/silexlabs/unifile/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/silexlabs/unifile/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/silexlabs/unifile/compare/v1.0.0...v1.1.0